+++ changelog.md
@@ -2,9 +2,16 @@
 Notable changes to the `cc.isr.vi` solution are documented in this file using the 
 [Keep a Changelog] style. The dates specified are in coordinated universal time (UTC).
 
-[8.1.9405]: https://www.github.com/atecoder/dn.vi.ivi
+[8.1.9407]: https://www.github.com/atecoder/dn.vi.ivi
 
 Current IVI Visa Compatibility: 8.0.1.9391
+
+## [8.1.9407] - 2025-10-03
+- Device.tsp Session Base Extensions
+  - Trim: Add trim script info extension method.
+  - Load: remove options from boolean arguments.
+  - Import: remove options from boolean arguments.
+  - Compile: Update the compile script: do not delete the script to allow compile and load over an existing script.
 
 ## [8.1.9405] - 2025-10-01
 - Device.TSP